Extract LoaderSpy to a separate file
